+++ src/ksp/src/KSPGraph.cpp
@@ -12,14 +12,14 @@ void KSPGraph::StartLoad() { clear();}
 void KSPGraph::EndLoad(void) {}
 void KSPGraph::clear(void) { Graph::clear();}
 
-void KSPGraph::AddData(ksp_edge_t  * edges,  int total_tuples, bool has_reverse_cost) {
-       for (int i = 0; i < total_tuples; i++) {
+void KSPGraph::AddData(ksp_edge_t  * edges,  UINT total_tuples, bool has_reverse_cost) {
+       for (UINT i = 0; i < total_tuples; i++) {
             insertTuple(edges[i].id, edges[i].source, edges[i].target,
             edges[i].cost, edges[i].reverse_cost, has_reverse_cost);
        }
 }
 
-void KSPGraph::insertTuple(int id, int source, int target,
+void KSPGraph::insertTuple(long id, long source, long target,
                 float cost, float reverse_cost, bool has_reverse_cost) {
     if (cost >= 0 || (reverse_cost>= 0 && has_reverse_cost)) {
        BaseVertex* sourcePos = getNewVertex(source);
